--- VideoMAEv2/__init__.py
@@ -1,1 +1,0 @@
-from .VideoMAEv2 import *
